[PATCH publishing] break out imageset label and move to subdirectory
