+++ src/Components/Publishing/Sections/ImageSetPreview/__tests__/ImageSetPreviewClassic.test.tsx
