+++ src/Components/Publishing/Sections/ImageSetPreview/index.ts
@@ -1,0 +1,3 @@
+import { ImageSetPreview } from "./ImageSetPreview"
+export default ImageSetPreview
+export * from "./ImageSetPreview"
